Unpause connectors on subscription
