+++ front/pages/api/stripe/webhook.ts
@@ -5,14 +5,15 @@ import type Stripe from "stripe";
 import { promisify } from "util";
 import { z } from "zod";
 
-import apiConfig from "@app/lib/api/config";
-import { getDataSources } from "@app/lib/api/data_sources";
 import {
   sendAdminSubscriptionPaymentFailedEmail,
   sendCancelSubscriptionEmail,
   sendReactivateSubscriptionEmail,
 } from "@app/lib/api/email";
-import { getMembers } from "@app/lib/api/workspace";
+import {
+  getMembers,
+  restoreWorkspaceAfterSubscription,
+} from "@app/lib/api/workspace";
 import { Authenticator } from "@app/lib/auth";
 import { startCreditFromProOneOffInvoice } from "@app/lib/credits/committed";
 import { grantFreeCreditsOnSubscriptionRenewal } from "@app/lib/credits/free";
@@ -34,7 +35,6 @@ import { countActiveSeatsInWorkspace } from "@app/lib/plans/usage/seats";
 import { WorkspaceModel } from "@app/lib/resources/storage/models/workspace";
 import { generateRandomModelSId } from "@app/lib/resources/string_ids";
 import { SubscriptionResource } from "@app/lib/resources/subscription_resource";
-import { TriggerResource } from "@app/lib/resources/trigger_resource";
 import { WorkspaceResource } from "@app/lib/resources/workspace_resource";
 import { ServerSideTracking } from "@app/lib/tracking/server";
 import { withTransaction } from "@app/lib/utils/sql_utils";
@@ -42,13 +42,10 @@ import { renderLightWorkspaceType } from "@app/lib/workspace";
 import logger from "@app/logger/logger";
 import { statsDClient } from "@app/logger/statsDClient";
 import { apiError, withLogging } from "@app/logger/withlogging";
-import {
-  launchScheduleWorkspaceScrubWorkflow,
-  terminateScheduleWorkspaceScrubWorkflow,
-} from "@app/temporal/scrub_workspace/client";
+import { launchScheduleWorkspaceScrubWorkflow } from "@app/temporal/scrub_workspace/client";
 import { launchWorkOSWorkspaceSubscriptionCreatedWorkflow } from "@app/temporal/workos_events_queue/client";
 import type { WithAPIErrorResponse } from "@app/types";
-import { assertNever, ConnectorsAPI, removeNulls } from "@app/types";
+import { assertNever } from "@app/types";
 
 export type GetResponseBody = {
   success: boolean;
@@ -281,7 +278,7 @@ async function handler(
                 subscriptionStartAt: now,
               });
             }
-            await onCancelScrub(
+            await restoreWorkspaceAfterSubscription(
               await Authenticator.internalAdminForWorkspace(workspace.sId)
             );
 
@@ -698,7 +695,7 @@ async function handler(
             );
             if (!endDate) {
               // Subscription is re-activated, so we need to unpause the connectors and re-enable triggers.
-              await onCancelScrub(auth);
+              await restoreWorkspaceAfterSubscription(auth);
 
               ServerSideTracking.trackSubscriptionReactivated({
                 workspace: renderLightWorkspaceType({
@@ -968,53 +965,6 @@ function _returnStripeApiError(
       message: `[Stripe Webhook][${event}] ${message}`,
     },
   });
-}
-
-/**
- * When a subscription is re-activated, we need to:
- * - Terminate the scheduled workspace scrub workflow (if any)
- * - Unpause all connectors
- * - Re-enable all triggers that point to non-archived agents
- */
-async function onCancelScrub(auth: Authenticator) {
-  const owner = auth.workspace();
-  if (!owner) {
-    throw new Error("Missing workspace on auth.");
-  }
-  const scrubCancelRes = await terminateScheduleWorkspaceScrubWorkflow({
-    workspaceId: owner.sId,
-  });
-  if (scrubCancelRes.isErr()) {
-    logger.error(
-      { stripeError: true, error: scrubCancelRes.error },
-      "Error terminating scrub workspace workflow."
-    );
-  }
-  const dataSources = await getDataSources(auth);
-  const connectorIds = removeNulls(dataSources.map((ds) => ds.connectorId));
-  const connectorsApi = new ConnectorsAPI(
-    apiConfig.getConnectorsAPIConfig(),
-    logger
-  );
-  for (const connectorId of connectorIds) {
-    const r = await connectorsApi.unpauseConnector(connectorId);
-    if (r.isErr()) {
-      logger.error(
-        { connectorId, stripeError: true, error: r.error },
-        "Error unpausing connector after subscription reactivation."
-      );
-    }
-  }
-
-  // Re-enable all triggers that point to non-archived agents
-  const enableTriggersRes = await TriggerResource.enableAllForWorkspace(auth);
-  if (enableTriggersRes.isErr()) {
-    logger.error(
-      { stripeError: true, error: enableTriggersRes.error },
-      "Error re-enabling workspace triggers on subscription reactivation"
-    );
-    // Don't throw error here - we want the function to continue even if trigger re-enabling fails
-  }
 }
 
 export default withLogging(handler);
